Update javaScript modules
- Fix  esLint issues
- Split JS into two modules. main.js handles step interaction. friends.js handles dynamic friend content.
